refactor:  move middlewares and routes folders inside main folder and formating existent files
