+++ src/main/adapters/index.ts
