+++ src/main/config/routes.ts
@@ -5,7 +5,7 @@ import { readdirSync } from 'fs'
 export default (app: Express): void => {
   const router = Router()
   app.use('/api', router)
-  readdirSync(`${__dirname}/routes`).map(async file => {
-    (await import(`${__dirname}/routes/${file}`)).default(router)
+  readdirSync(`${__dirname}/../routes`).map(async file => {
+    (await import(`${__dirname}/../routes/${file}`)).default(router)
   })
 }
