+++ src/main/factories/register.ts
@@ -1,6 +1,6 @@
 import { RegisterUserController } from '@/web-controllers'
-import { RegisterUserOnMailingList } from '@/usecases/register-user-on-maling-list'
-import { InMemoryUserRepository } from '@test/usecases/register-user-on-maling-list/repository'
+import { RegisterUserOnMailingList } from '@/usecases/register-user-on-mailing-list'
+import { InMemoryUserRepository } from '@/usecases/register-user-on-mailing-list/repository'
 
 export const makeRegisterUserController = (): RegisterUserController => {
   const inMemoryUserRepository = new InMemoryUserRepository([])
