+++ src/main/middleware/content-type.ts
