+++ src/main/routes/register-route.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 import { makeRegisterUserController } from '@/main/factories'
-import { adaptRoute } from '@/main/config/adapters'
+import { adaptRoute } from '@/main/adapters'
 
 export default (router: Router):void => {
   router.post('/register', adaptRoute(makeRegisterUserController()))
